🎉 library setup  - added testing and publishing workflows  - added library code to src folder  - added dependencies  - added required script
