+++ .changeset/bright-fans-clap.md
@@ -1,0 +1,5 @@
+---
+"textcon": patch
+---
+
+setup repo - add base function and setup publising pipeline
