+++ package.json
@@ -1,0 +1,47 @@
+{
+  "name": "textcon",
+  "version": "0.0.0",
+  "description": "A react state management library that uses context and provide selector support for fine grainded reactivity",
+  "main": "dist/index.cjs.js",
+  "module": "dist/index.es.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "test": "vitest",
+    "lint": "tsc",
+    "test:all": "pnpm run lint && pnpm run test",
+    "build": "vite build"
+  },
+  "keywords": [
+    "react",
+    "react-context",
+    "selectors",
+    "state-management"
+  ],
+  "author": "Abdullah Adeel (contact.abdadeel@gmail.com)",
+  "license": "ISC",
+  "devDependencies": {
+    "@changesets/cli": "^2.26.1",
+    "@testing-library/react": "^14.0.0",
+    "@types/jsdom": "^21.1.1",
+    "@types/lodash.isequal": "^4.5.6",
+    "@types/node": "^20.2.5",
+    "@types/react": "^18.2.7",
+    "@typescript-eslint/eslint-plugin": "^5.59.7",
+    "@typescript-eslint/parser": "^5.59.7",
+    "@vitejs/plugin-react": "^4.0.0",
+    "dts-bundle-generator": "^8.0.1",
+    "eslint": "^8.41.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jsdom": "^22.1.0",
+    "prettier": "^2.8.8",
+    "typescript": "^5.0.4",
+    "vite": "^4.3.9",
+    "vite-plugin-dts": "^2.3.0",
+    "vitest": "^0.31.1"
+  },
+  "dependencies": {
+    "lodash.isequal": "^4.5.0",
+    "react": "^18.2.0"
+  }
+}
