+++ vite.config.ts
@@ -1,0 +1,27 @@
+import path from "path";
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import dts from "vite-plugin-dts";
+
+export default defineConfig({
+  base: "./",
+  plugins: [dts(), react()],
+  build: {
+    sourcemap: true,
+    lib: {
+      entry: path.resolve(__dirname, "src/index.ts"),
+      name: "textcon",
+      formats: ["es", "cjs", "umd", "iife"],
+      fileName: (format) => `index.${format}.js`,
+    },
+    rollupOptions: {
+      external: ["react", "react-dom"],
+      output: {
+        globals: {
+          react: "React",
+          "react-dom": "ReactDOM",
+        },
+      },
+    },
+  },
+});
