+++ vitest.config.ts
@@ -1,0 +1,21 @@
+/// <reference types="vitest" />
+/// <reference types="vite/client" />
+
+import { defineConfig, mergeConfig } from "vitest/config";
+
+import viteConfig from "./vite.config";
+
+export default mergeConfig(
+  viteConfig,
+  defineConfig({
+    test: {
+      environment: "jsdom",
+      testTimeout: 30_000,
+      hookTimeout: 30_000,
+      include: [
+        "**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}",
+        "../../packages/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}",
+      ],
+    },
+  })
+);
